refactor: move Queue to dispatcher package
